# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorCode": {
        "description": "Stable API error codes consumed by the web client.",
        "enum": [
          "VALIDATION_ERROR",
          "EMPTY_FILE",
          "FILE_TOO_LARGE",
          "UNSUPPORTED_MEDIA_TYPE",
          "INVALID_ENCODING",
          "MALFORMED_FDX",
          "MALFORMED_PDF",
          "INVALID_NATIVE_DOCUMENT",
          "PROCESS_RESTARTED",
          "UNAUTHENTICATED",
          "AUTHENTICATION_UNAVAILABLE",
          "ACCOUNT_BINDING_CONFLICT",
          "DOCUMENT_NOT_FOUND",
          "TASK_NOT_FOUND",
          "NOT_FOUND",
          "FORBIDDEN",
          "CONFLICT",
          "DOCUMENT_CONTENT_MISSING",
          "CONTRACT_MISMATCH",
          "INTERNAL_ERROR"
        ],
        "title": "ApiErrorCode",
        "type": "string"
      },
      "ApiErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorDetail",
        "type": "object"
      },
      "ApiErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorEnvelope",
        "type": "object"
      },
...
      "DocumentItem": {
        "description": "Summary item for document list view.",
        "properties": {
          "arc_distance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arc Distance"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "current_version_id": {
            "title": "Current Version Id",
            "type": "integer"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "expected_version": {
            "title": "Expected Version",
            "type": "integer"
          },
          "gcs_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gcs Uri"
          },
          "is_saved": {
            "title": "Is Saved",
            "type": "boolean"
          },
          "owner_email": {
            "title": "Owner Email",
            "type": "string"
          },
          "owner_user_id": {
            "title": "Owner User Id",
            "type": "string"
          },
          "shared_count": {
            "title": "Shared Count",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "version_id": {
            "title": "Version Id",
            "type": "integer"
          }
        },
        "required": [
          "document_id",
          "owner_user_id",
          "title",
          "current_version_id",
          "version_id",
          "expected_version",
          "is_saved",
          "created_at",
          "updated_at",
          "owner_email",
          "shared_count"
        ],
        "title": "DocumentItem",
        "type": "object"
      },
      "DocumentListResponse": {
        "description": "Paginated document list response.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "DocumentListResponse",
        "type": "object"
      },
...
      "TaskAcceptedResponse": {
        "properties": {
          "status": {
            "default": "queued",
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "task_type": {
            "title": "Task Type",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "task_type"
        ],
        "title": "TaskAcceptedResponse",
        "type": "object"
      },
...
      "ValencePatternItem": {
        "description": "One configured valence-arc pattern safe to expose to the UI.",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pattern_id": {
            "title": "Pattern Id",
            "type": "string"
          }
        },
        "required": [
          "pattern_id",
          "name",
          "description"
        ],
        "title": "ValencePatternItem",
        "type": "object"
      },
      "ValencePatternListResponse": {
        "description": "Configured Valence pattern catalogue in its canonical config order.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ValencePatternItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ValencePatternListResponse",
        "type": "object"
      },
      "ValenceSimilarityItem": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pattern_id": {
            "title": "Pattern Id",
            "type": "string"
          },
          "percentage": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Percentage",
            "type": "number"
          }
        },
        "required": [
          "pattern_id",
          "name",
          "description",
          "percentage"
        ],
        "title": "ValenceSimilarityItem",
        "type": "object"
      },
      "ValenceSimilarityResponse": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ValenceSimilarityItem"
            },
            "title": "Items",
            "type": "array"
          },
          "version_id": {
            "title": "Version Id",
            "type": "integer"
          }
        },
        "required": [
          "document_id",
          "version_id",
          "items"
        ],
        "title": "ValenceSimilarityResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Intelligent screenplay analysis and narrative arc workbench API",
    "title": "NARRAVANT API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/documents": {
      "get": {
        "description": "List screenplays with optional filtering and narrative arc similarity matching.",
        "operationId": "list_documents_api_v1_documents_get",
        "parameters": [
          {
            "description": "Search term for title",
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search term for title",
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Repeated field:direction server-side sort terms",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeated field:direction server-side sort terms",
              "title": "Sort"
            }
          },
          {
            "description": "Document ID to find similar narrative arcs",
            "in": "query",
            "name": "similar_to_arc_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Document ID to find similar narrative arcs",
              "title": "Similar To Arc Id"
            }
          },
          {
            "description": "Configured Valence pattern ID to find similar narrative arcs; mutually exclusive with similar_to_arc_id",
            "in": "query",
            "name": "arc_pattern_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Configured Valence pattern ID to find similar narrative arcs; mutually exclusive with similar_to_arc_id",
              "title": "Arc Pattern Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Documents",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/documents/{doc_id}/emotion-arc/reanalyze": {
      "post": {
        "description": "Queue an owner-only reanalysis that returns an unsaved emotion-arc draft by SSE.",
        "operationId": "reanalyze_emotion_arc_api_v1_documents__doc_id__emotion_arc_reanalyze_post",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Reanalyze Emotion Arc",
        "tags": [
          "emotion-arc"
        ]
      }
    },
    "/api/v1/documents/{doc_id}/emotion-arc/similarities": {
      "get": {
        "description": "Compare the selected document's completed valence arc to all configured Valence patterns.",
        "operationId": "get_emotion_arc_similarities_api_v1_documents__doc_id__emotion_arc_similarities_get",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValenceSimilarityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Emotion Arc Similarities",
        "tags": [
          "emotion-arc"
        ]
      }
    },
...
    "/api/v1/emotion-arc/patterns": {
      "get": {
        "description": "Return labels from config.yaml; curve vectors remain server-side search inputs.",
        "operationId": "get_valence_patterns_api_v1_emotion_arc_patterns_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValencePatternListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Valence Patterns",
        "tags": [
          "emotion-arc"
        ]
      }
    },
...
  }
}